# api/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "auth_internal_admin_roles.CreateRoleParams": {
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "auth_internal_admin_users.GetUserResponse": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "updated_at": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "events.EventResponse": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "ip_address": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "user_agent": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "events.ListEventsResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/events.EventResponse"
            },
            "type": "array"
          },
          "next_cursor": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "refreshtokens.ListRefreshTokensResponse": {
        "properties": {
          "next_cursor": {
            "type": "string"
          },
          "refresh_tokens": {
            "items": {
              "$ref": "#/components/schemas/refreshtokens.RefreshTokenResponse"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "refreshtokens.RefreshTokenResponse": {
        "properties": {
          "expires_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "ip_address": {
            "type": "string"
          },
          "issued_at": {
            "type": "string"
          },
          "revoked_at": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "user_agent": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "users.ListUsersResponse": {
        "properties": {
          "next_cursor": {
            "type": "string"
          },
          "users": {
            "items": {
              "$ref": "#/components/schemas/users.UserWithMetadata"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "users.UpdateUserRoleBody": {
        "properties": {
          "role": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "users.UserWithMetadata": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "updated_at": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Type \"Bearer\" followed by a space and JWT token.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "api for auth.hisbaan.com",
    "title": "auth",
    "version": "0.0.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/admin/events/users/{userId}": {
      "get": {
        "description": "Returns a paginated list of audit events for a user (admin only)",
        "parameters": [
          {
            "description": "User ID",
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Pagination cursor",
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum number of results (max 100)",
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/events.ListEventsResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List user events",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/refresh-tokens/users/{userId}": {
      "get": {
        "description": "Returns a paginated list of refresh tokens for a user (admin only)",
        "parameters": [
          {
            "description": "User ID",
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Pagination cursor",
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum number of results (max 100)",
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/refreshtokens.ListRefreshTokensResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List user refresh tokens",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/roles": {
      "post": {
        "description": "Creates a new role (admin only)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth_internal_admin_roles.CreateRoleParams"
              }
            }
          },
          "description": "Role details",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create role",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/roles/{name}": {
      "delete": {
        "description": "Deletes a role (admin only)",
        "parameters": [
          {
            "description": "Role name",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete role",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "description": "Returns a paginated list of users (admin only)",
        "parameters": [
          {
            "description": "Pagination cursor",
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum number of results (max 100)",
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/users.ListUsersResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List users",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{id}": {
      "get": {
        "description": "Returns a user by their ID (admin only)",
        "parameters": [
          {
            "description": "User ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth_internal_admin_users.GetUserResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{userId}/roles": {
      "post": {
        "description": "Assigns a role to a user (admin only)",
        "parameters": [
          {
            "description": "User ID",
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users.UpdateUserRoleBody"
              }
            }
          },
          "description": "User role assignment",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Add user role",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{userId}/roles/{role}": {
      "delete": {
        "description": "Removes a role from a user (admin only)",
        "parameters": [
          {
            "description": "User ID",
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Role name",
            "in": "path",
            "name": "role",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Remove user role",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}